Allow decimal print times and display seconds for short durations
- Change print time input to allow decimals (step=0.1)
- Display as seconds when under 1 minute (e.g., "30 sec")
- Update API routes to use parseFloat instead of parseInt
- Fix POST route to actually save print time and ink usage

+++ src/app/api/samples/[id]/route.ts
@@ -84,7 +84,7 @@ export async function PATCH(
       const inkUsageStr = formData.get('ink_usage_ml') as string | null
       const onedriveUrlStr = formData.get('onedrive_folder_url') as string | null
 
-      print_time_minutes = printTimeStr ? parseInt(printTimeStr) : null
+      print_time_minutes = printTimeStr ? parseFloat(printTimeStr) : null
       ink_usage_ml = inkUsageStr ? parseFloat(inkUsageStr) : null
       onedrive_folder_url = onedriveUrlStr || null
